requirements.txt
mcp>=1.0.0
httpx>=0.25.0
asyncpg>=0.29.0
sqlalchemy>=2.0.0
alembic>=1.12.0
scikit-learn>=1.3.0
pandas>=2.1.0
numpy>=1.25.0
python-dotenv>=1.0.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0
jupyter>=1.0.0
matplotlib>=3.7.0
seaborn>=0.12.0